telegram_bot/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
asyncio==4.0.0
attrs==25.4.0
beautifulsoup4==4.14.2
billiard==4.2.4
cachetools==6.2.4
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
deprecation==2.1.0
exceptiongroup==1.3.1
fastapi==0.127.0
frozenlist==1.8.0
fsspec==2025.12.0
greenlet==3.3.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
kombu==5.6.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.0
numpy==2.2.6
packaging==25.0
pandas==2.3.3
postgrest==2.27.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.3.1
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-multipart==0.0.21
python-telegram-bot
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.10.1
realtime==2.27.0
redis==7.1.0
redis[hiredis]==7.1.0
requests==2.32.5
rich==14.2.0
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.45
starlette==0.50.0
storage3==2.27.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.0
supabase-auth==2.27.0
supabase-functions==2.27.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.2
uvicorn==0.40.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
yarl==1.22.0
